// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  NONBINARY: 'NONBINARY'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const SupportType = {
  NURSE: 'NURSE',
  ALLIED_HEALTH: 'ALLIED_HEALTH',
  PERSONAL_CARE: 'PERSONAL_CARE',
  DAILY_LIVING: 'DAILY_LIVING',
  COMMUNITY: 'COMMUNITY',
  OTHER: 'OTHER'
} as const

export type SupportType = (typeof SupportType)[keyof typeof SupportType]


export const StaffStatus = {
  ONBOARDING: 'ONBOARDING',
  ACTIVE: 'ACTIVE',
  TERMINATED: 'TERMINATED'
} as const

export type StaffStatus = (typeof StaffStatus)[keyof typeof StaffStatus]


export const RepeatitiveFrequency = {
  DAILY: 'DAILY',
  WEEKLY: 'WEEKLY',
  MONTHLY: 'MONTHLY',
  YEARLY: 'YEARLY'
} as const

export type RepeatitiveFrequency = (typeof RepeatitiveFrequency)[keyof typeof RepeatitiveFrequency]